feat: add react setup and update tests
