+++ README.md
@@ -1,1 +1,20 @@
-# productioncontrol
+# Production Control
+
+Este proyecto implementa un tablero Kanban sincronizado con Firebase para gestionar pedidos.
+
+## Instalación
+
+```bash
+npm install
+```
+
+## Comandos
+
+- `npm start` inicia un servidor de desarrollo y abre la aplicación en el navegador.
+- `npm run build` ejecuta el paso de construcción (actualmente no implementado).
+- `npm test` lanza las pruebas con Jest.
+
+## Dependencias Principales
+
+- **React** y **React DOM** para interfaces de usuario.
+- **Jest** para pruebas automatizadas.
