+++ package.json
@@ -4,10 +4,20 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-  "test": "jest"
+    "start": "http-server -o",
+    "build": "echo 'No build step defined'",
+    "test": "NODE_OPTIONS=--experimental-vm-modules jest"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "type": "commonjs"
+  "type": "module",
+  "dependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
+  },
+  "devDependencies": {
+    "jest": "^29.6.1",
+    "http-server": "^14.1.1"
+  }
 }
