+++ test/kanban.test.js
@@ -1,19 +1,19 @@
-const { getColumnColorByClientes } = require('../kanban');
+import { getColumnColorByClientes, stringToColor } from '../kanbanUtils.js';
 
 describe('getColumnColorByClientes', () => {
-  test('debe retornar verde si todos los pedidos son del mismo cliente', () => {
+  test('debe retornar el color del cliente cuando todos son iguales', () => {
     const pedidos = [
       { cliente: 'Cliente A' },
       { cliente: 'Cliente A' }
     ];
-    expect(getColumnColorByClientes(pedidos)).toBe('lightgreen');
+    expect(getColumnColorByClientes(pedidos)).toBe(stringToColor('Cliente A', 90, 96));
   });
 
-  test('debe retornar gris si hay múltiples clientes', () => {
+  test('debe retornar el color del cliente más frecuente o el primero', () => {
     const pedidos = [
       { cliente: 'Cliente A' },
       { cliente: 'Cliente B' }
     ];
-    expect(getColumnColorByClientes(pedidos)).toBe('lightgray');
+    expect(getColumnColorByClientes(pedidos)).toBe(stringToColor('Cliente A', 90, 96));
   });
 });
